+++ nautobot_secrets_providers/providers/__init__.py
@@ -2,5 +2,11 @@
 
 from .aws import AWSSecretsManagerSecretsProvider
 from .hashicorp import HashiCorpVaultSecretsProvider
+from .thycotic import ThycoticSecretServerSecretsProviderId, ThycoticSecretServerSecretsProviderPath
 
-__all__ = (AWSSecretsManagerSecretsProvider, HashiCorpVaultSecretsProvider)  # pylint: disable=invalid-all-object
+__all__ = (  # type: ignore
+    AWSSecretsManagerSecretsProvider,  # pylint: disable=invalid-all-object
+    HashiCorpVaultSecretsProvider,  # pylint: disable=invalid-all-object
+    ThycoticSecretServerSecretsProviderId,  # pylint: disable=invalid-all-object
+    ThycoticSecretServerSecretsProviderPath,  # pylint: disable=invalid-all-object
+)