+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Nautobot Secrets Providers Changelog
 
+## v1.1.0 (2022-03-10)
+
+### Added
+
+- [#21](https://github.com/nautobot/nautobot-plugin-secrets-providers/issues/21) Add Thycotic Secret Server plugin
+  **Requires Python 3.7 or greater**
+
 ## v1.0.1 (2022-01-06)
 
 ### Fixed